minimized the tailor dashboard sidebar codebase by creating components
